Remove NSAttributedStringKey+Conversion
