+++ WordPress/Classes/ViewRelated/Reader/Cards/ReaderBlockedSiteCell.swift
@@ -39,3 +39,18 @@ open class ReaderBlockedSiteCell: UITableViewCell {
     }
 
 }
+
+private extension NSAttributedString.Key {
+    /// Converts a collection of NSAttributedString Attributes, with 'String' instances as 'Keys', into an equivalent
+    /// collection that uses the new 'NSAttributedStringKey' enum as keys.
+    ///
+    static func convertFromRaw(attributes: [String: Any]) -> [NSAttributedString.Key: Any] {
+        var output = [NSAttributedString.Key: Any]()
+        for (key, value) in attributes {
+            let wrappedKey = NSAttributedString.Key(key)
+            output[wrappedKey] = value
+        }
+
+        return output
+    }
+}
