+++ WordPress/Classes/ViewRelated/Reader/Style/WPStyleGuide+ReaderComments.swift
@@ -33,3 +33,17 @@ extension WPStyleGuide {
         static let switchInProgressTintColor = UIAppColor.primary
     }
 }
+
+private extension NSAttributedString.Key {
+    /// Converts a collection of NSAttributedString Attributes, with 'NSAttributedStringKey' instances as 'Keys', into an
+    /// equivalent collection that uses regular 'String' instances as keys.
+    ///
+    static func convertToRaw(attributes: [NSAttributedString.Key: Any]) -> [String: Any] {
+        var output = [String: Any]()
+        for (key, value) in attributes {
+            output[key.rawValue] = value
+        }
+
+        return output
+    }
+}
